Exceptions adicionada

+++ src/main/java/com/malbano/rural/model/dto/AcumuloDTO.java
@@ -15,7 +15,7 @@ public class AcumuloDTO {
 
     private String nomeProduto;
     private String anoEmissao;
-    private Double vlCusteioTotal;
+    private BigDecimal vlCusteioTotal;
 
 
     public static AcumuloDTO create(DadosDTO dados) {
